refactor(`@prices`): changed directory configuration
