+++ prices/service-entrypoint/src/bootstrap/index.ts
@@ -1,0 +1,1 @@
+export * from './bootstrap.js'
