+++ prices/service-entrypoint/src/cron/index.ts
@@ -1,0 +1,1 @@
+export * from './cron.js'
