+++ prices/service-entrypoint/src/entrypoint/index.ts
@@ -1,0 +1,1 @@
+export * from './entrypoint.js'
