+++ prices/service-entrypoint/src/index.ts
@@ -1,1 +1,1 @@
-import './bootstrap.js'
+import './entrypoint/index.js'
